Factor out GuideRenderer interface
So that we can strategize the way that the rendering gets done
(starting with just one option, local asciidoc).
