+++ sagan-common/src/main/java/sagan/guides/support/UnderstandingDocs.java
@@ -1,18 +1,21 @@
 package sagan.guides.support;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
-import sagan.guides.*;
+import sagan.guides.ContentProvider;
+import sagan.guides.UnderstandingDoc;
+import sagan.guides.UnderstandingMetadata;
 import sagan.support.ResourceNotFoundException;
 import sagan.support.github.RepoContent;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClientException;
 
@@ -20,22 +23,23 @@ import org.springframework.web.client.RestClientException;
  * Repository implementation providing data access services for understanding docs.
  */
 @Component
-public class UnderstandingDocs implements DocRepository<UnderstandingDoc, UnderstandingMetadata>, ContentProvider<UnderstandingDoc> {
+public class UnderstandingDocs implements DocRepository<UnderstandingDoc, UnderstandingMetadata>,
+        ContentProvider<UnderstandingDoc> {
 
     private static final Log log = LogFactory.getLog(UnderstandingDocs.class);
 
     public static final String CACHE_NAME = "cache.understanding";
-    public static final Class CACHE_TYPE = UnderstandingDoc.class;
+    public static final Class<?> CACHE_TYPE = UnderstandingDoc.class;
     public static final String CACHE_TTL = "${cache.docs.timetolive:0}"; // never expires
 
     private static final String CONTENT_PATH = "/repos/%s/%s/contents/%s/README.md";
-    private static final String SIDEBAR_PATH = "/repos/%s/%s/contents/%s/SIDEBAR.md";
 
     private final GuideOrganization org;
     private final String repoName;
 
     @Autowired
-    public UnderstandingDocs(GuideOrganization org, @Value("${github.repo.understanding:understanding}") String repoName) {
+    public UnderstandingDocs(GuideOrganization org,
+                             @Value("${github.repo.understanding:understanding}") String repoName) {
         this.org = org;
         this.repoName = repoName;
     }
@@ -45,6 +49,7 @@ public class UnderstandingDocs implements DocRepository<UnderstandingDoc, Unders
      *
      * @throws ResourceNotFoundException if no content exists for the subject.
      */
+    @Override
     @Cacheable(value = CACHE_NAME)
     public UnderstandingDoc find(String subject) {
         UnderstandingDoc doc = new UnderstandingDoc(subject);
